+++ protected/humhub/modules/content/widgets/WallCreateContentForm.php
@@ -8,16 +8,16 @@
 
 namespace humhub\modules\content\widgets;
 
+use humhub\components\Widget;
+use humhub\modules\content\components\ContentActiveRecord;
+use humhub\modules\content\components\ContentContainerActiveRecord;
+use humhub\modules\content\models\Content;
 use humhub\modules\content\permissions\CreatePublicContent;
+use humhub\modules\space\models\Space;
 use humhub\modules\stream\actions\StreamEntryResponse;
 use humhub\modules\topic\models\Topic;
 use humhub\modules\ui\form\widgets\ActiveForm;
-use humhub\components\Widget;
 use humhub\modules\user\models\User;
-use humhub\modules\space\models\Space;
-use humhub\modules\content\models\Content;
-use humhub\modules\content\components\ContentContainerActiveRecord;
-use humhub\modules\content\components\ContentActiveRecord;
 use Yii;
 use yii\web\HttpException;
 
@@ -131,7 +131,7 @@ abstract class WallCreateContentForm extends Widget
 
         if ($record->save()) {
             $topics = Yii::$app->request->post('postTopicInput');
-            if (!empty($topics)) {
+            if (!empty($topics) && Topic::isAllowedToCreate($contentContainer)) {
                 Topic::attach($record->content, $topics);
             }