+++ CHANGELOG.md
@@ -62,6 +62,7 @@ HumHub Changelog
 - Enh #7265: Profile "About" page: don't display the menu if only one entry
 - Enh #7269: Remove desktop notifications
 - Enh #5423: Display password rules in form field hint
+- Enh #7223: Add ability to convert Topic to global
 
 1.16.3 (Unreleased)
 --------------------------